1. FlipBookComponent 설계 2. FlipBook 설계 3. Sprite 설계 4. AssetMgr CreateEngineSprite 함수 구현 5. MeshRender Render 함수FlipBookComponent 분기 추가 6. HLSL PixelShader 함수 Std2D 함수 FlipBook 사용 여부 처리

+++ Project/Engine/CTaskMgr.cpp
@@ -37,7 +37,7 @@ void CTaskMgr::ExecuteTask()
 		case TASK_TYPE::CREATE_OBJECT:
 		{
 			CLevel* pCurLevel = CLevelMgr::GetInst()->GetCurrentLevel();
-			int Layer = m_vecTask[i].Param_0;
+			int Layer = (int)m_vecTask[i].Param_0;
 			CGameObject* pObj = (CGameObject*)m_vecTask[i].Param_1;
 			pCurLevel->AddObject(Layer, pObj);
 		}
